FastAPI_app/requirements.txt
sqlalchemy
fastapi
uvicorn
python-dotenv
pydantic
psycopg2